Still nothing major; feeling bad.
With how terrible I was feeling, I managed to get the inputs completed in terms of the verification, and nothing else. I couldn't even properly float the elements to the left of the page. Something was over me today and it took all my energy away, so hopefully I'll get that back with another day.